Remove support for NewType typedef annotation
Summary:
I have no idea why I even implemented this in the first place.
I can't think of any applications where this might be useful.

The feature is completely unused.
The feature was not supported in the legacy generator (type aliases were used always):
https://www.internalfb.com/code/fbsource/[bbea5a1ec838]/fbcode/thrift/compiler/generate/t_go_generator.cc?lines=971

So I think we can safely kill it.

Reviewed By: leoleovich

Differential Revision: D52568785

fbshipit-source-id: 929d6f36f70e1f9608484ce9983a6f03cb44aef6