RFC: [thrift][folly_dynamic] @cpp.Adapter compatibility with to_dynamic and from_dynamic
Summary: This diff is a proposal for adding support for Thrift's `cpp.Adapter` annotation to the new `folly_dynamic` library, specifically to `to_dynamic` and `from_dynamic functions`. This would allow for more seamless migration between old and new API, making it easier to use Thrift data structures with Adapter.

Reviewed By: thedavekwon

Differential Revision: D52515956

fbshipit-source-id: 78d1566fd0a0e5bb2a53d6996f5a566470bbc3c5