thrift/rust: easy: unit test serde deserialization errors
Summary: Adding a unit test to secure the current behavior of deserialization errors via serde and the state of its compatibility with simplejson. Whilst this might or might not be optimal, any changes to this would be a breaking change of compatibility, so having a unit test should make sure any such changes happen in a controlled way.

Reviewed By: dtolnay

Differential Revision: D52787569

fbshipit-source-id: 83e37c7cbc04cf0e36925fe21322a42a7527d401