Use optional second component of namespace to set module name
Summary:
Previously, the `namespace rust` for a single-file Thrift library had to be a single component: `namespace rust cratename`, while for a multi-file Thrift library it had to be 2 components: `namespace rust cratename.something`, where `something` could be an arbitrary identifier that was ignored. Notably, it would **not** change the name of the modules in the generated Rust library. The module for filename.thrift would be `cratename::filename`, not `cratename::something`.

This diff enables using `namespace rust cratename.modulename` to control the module names in a multi-file Thrift library. This is useful because many Thrift files in fbcode have CamelCase names, which are not idiomatic as a Rust module name.

Also, this diff makes the second namespace component optional for multi-file Thrift libraries. If you use `namespace rust cratename.modulename`, the generated code becomes accessible from Rust as `cratename::modulename`. If you use `namespace rust cratename`, the module name is taken from the Thrift source file's name, so `cratename::filename` for the module corresponding to filename.thrift.

Representative example:

https://www.internalfb.com/code/fbsource/[08760c0afd25e2fbde8d5e987a8c71cb03eb4098]/fbcode/conveyor/artifact_content/if/TARGETS?lines=5-7%2C9%2C11%2C13%2C16-18%2C44%2C50

https://www.internalfb.com/code/fbsource/[08760c0afd25e2fbde8d5e987a8c71cb03eb4098]/fbcode/conveyor/artifact_content/if/cloud_ami.thrift?lines=3-6

In this file we would want to add `namespace rust conveyor_artifact_content`.

It would be unfriendly and inconsistent with every other language to require `namespace rust conveyor_artifact_content.cloud_ami`.

Reviewed By: diliop

Differential Revision: D52822292

fbshipit-source-id: c53d7486ecaacd663bf9e3cd935e179e57cd00dd