Register underlying struct in TypeRegistry
Summary:
For Adapted Structs, the struct name refers to the adapted type. The underlying type should be registered instead.

Facebook :
Post : https://fb.workplace.com/groups/thriftusers/permalink/1334592907250118/

**Re: Skip registering adapted struct vs Registering underlying structs :**

Generated Patch Structs are adapted and the `underlyingName` is provided so that the underlying thrift type is generated with a different name instead of being generated in a nested namespace. The URI refers to the underlying struct.

```
package "facebook.com/my_package"
patch.GeneratePatch
struct Foo {}
```
In above eg, `facebook.com/my_package/FooPatch` refers to the `FooPatchStruct` and the adaptedType is named `FooPatch`.

ICSP stores underlying patch structs `<struct>PatchStruct` in `AnyStruct` using the URI of the underlying type. And they use `TypeRegistry` to store/load the Patch structs. **Changing the URI of underlying type or not registering underlying structs would break their builds.**

I believe the original intention was to always register the underlying type as the template uses `struct:cpp_underlying_name` The bug only shows up when there is an extra namespace.

For patch, both adapted type and underlying type are generated in the same namespace and when registering the underlying type is registered as expected.

Reviewed By: thedavekwon

Differential Revision: D52863003

fbshipit-source-id: a9196465b228a3aaaf95be57787a637e50faf002