Move enum aliasing logic to mutator
Summary:
This logic makes it possible to refer to enum values using aliases of the containing enum, but its placement in the parser creates order dependence between the enum and its alias. Move it to the mutator so the true type always resolves.
This requires keeping track of the program that a given identifier was used in, in case the identifier needs to be scoped during resolution. (This happens during a fairly contrived corner-case, captured as `alias_enum_in_external_const`, that occurs exactly once in our codebase).

Reviewed By: aristidisp

Differential Revision: D53011766

fbshipit-source-id: ac7eb9f69c96a0da6e3436ddd88add1d208a27bc