Fix package and remove deprecated annotations in testset
Summary: The package name in `thrift/test/testset/golden/testset.thrift` is wrong because we don't control the `apache.org` domain. Replace it with `facebook.com` and remove uses of the deprecated `thrift.uri` annotations which are no longer needed after fixing the package name. These constitute more than 90% uses of this annotation in fbcode.

Reviewed By: iahs

Differential Revision: D53269804

fbshipit-source-id: 7491a5c8935bfebba24166467eb1f4a7a5881e33