Fix crate name for dependency's subcrates when dependency's name is keyword
Summary:
Similar fix to D53270825 that I need for unblocking D53246158.

Previously, the Rust Thrift code generator might be told that the generated code is supposed to depend on a dependency called `if_` in 2 **different** situations:

1. If the dependency's target is really called `if`, and that has been mangled to `if_`. In this case the corresponding types crate is called **`if_types`**.

2. If the dependency's target is really called `if_` and no mangling was involved. The types crate is called **`if__types`** because it was created by concatenating `_types` onto the original name.

This diff fixes Thrift-related macros to be able to distinguish the `if` vs `if_` case and properly name the types subcrate in both.

Reviewed By: shayne-fletcher

Differential Revision: D53283225

fbshipit-source-id: 861b30f7fc7e25f725647ebd47d8620385b14981