Export a more generous API from types, clients, and services subcrates
Summary:
I expect this diff to significantly help with the Thrift crate redesign described in D53302058.

By making more of the API of the old `-rust-types` subcrate accessible through the `-rust-clients` and `-rust-services` public APIs, in a way that loosely follows the old `-rust` "main" crate, I can more easily refactor existing users of `-rust` to use one of `-rust-clients` or `-rust-services` as appropriate. Many times this will need only a single line change in TARGETS.

This diff prevents having to add separate `-rust-clients` (or `-rust-services`) and `-rust-types` dependencies for crates that used to depend on just a single `-rust` target, and mucking with imports in Rust source files to import from the right one of the two.

Reviewed By: shayne-fletcher

Differential Revision: D53338214

fbshipit-source-id: 0273c9764159c63120fb97679eec8ce44d165449