Soft deprecate `dependencies` modules
Summary:
See discussion in D53316562: I plan to burn down the usage of Thrift-generated `::dependencies::` by handwritten code in fbcode. Those modules only exist to support service inheritance in generated code across crates. (The generated code for a service needs to refer to the entire chain of super-services, including across dependencies, and it does that using these re-exports.)

This diff adds a comment above the re-exports identifying the correct Buck target and crate name to use instead. This is valuable to me in codemodding the existing uses. What I can do is locally change the comment into a real `#[deprecated = "..."]` attribute, then find all occurrences of `::dependencies::` in Rust files throughout the repo, and compile those targets with `-Ddeprecated -Awarnings` in rustflags. This will print out every import that needs to be updated and every dependency to use.

Reviewed By: emersonford

Differential Revision: D53374511

fbshipit-source-id: 51237f0009f707b1a73db67e3dcc63d14202d3a9