fix make_unique for cython 3.0.8
Summary:
make_unique sig changes in cython 3.0.8 so that it has `except+` which causes exception handling code to be generated around its usage. Which breaks all the places we use it because of move semantics.

This seems the easiest solution that doesn't involve a whole bunch of restructuring of code in the code generator.

Reviewed By: itamaro

Differential Revision: D53442598

fbshipit-source-id: e7c1117de7db6defb643c38c5df4e41690f56a3b