fix serialize-parity issue with constructed optional field
Summary:
When a thrift-python object is deserialized, and it has an optional field, the field's PyObject* data tuple is set to `Py_None`, without modifying the field's isset bit (leaving it unset). https://fburl.com/code/0oyklv9w8

Thrift-python C-API currently calls `setStructField` with `Py_None`, which incorrectly sets the `isset` bit. Although this yields `foo.bar is None` in python code as desired, it does not conform to object state expected after deserialization, causing an error on serialization. Due to the isset bit, the serializer expects a value (e.g., `bytes`) but then has a type error when it encounters the `None`.

Reviewed By: Filip-F

Differential Revision: D53452232

fbshipit-source-id: d7e853b494032e1cfdd7cb026d7d10c5ed8041f8