regenerate sbe flyweights
Summary:
Previously, we need to use gradle to compile `sbe-all` before we can generate flyweights from RpcMetadata.xml

in this diff, we use sbe-all imported from third-party-java to achieve the same. We updated RpcMetadata.xml on how to run this script

Please note that previously, we uses snapshot version of SBE 1.31.0-SNAPSHOT and in this diff, we uses the latest release version SBE 1.30.0, this explains minor differences in generated code.

We would like to eventually generate a c++ library from RpcMetadata.xml using buck, but we leave that to later as part of the production effort, for now, this is suffice.

Reviewed By: robertroeser

Differential Revision: D53976780

fbshipit-source-id: f4a375036f66c70113dc0414275199a17b12d866