Make MapPatch runtime library work with incomplete type
Summary:
Currently if there is a circular dependency between thrift structs, we have to mark the struct as AssignOnlyPatch in order to use thrift patch, since the existing patch implementation can not handle incomplete type.

However, AssignOnlyPatch not only introduces extra complexity (we need special code to handle it), but also doesn't work with other patch features (e.g., it ignores other operations and only recognize assign operation, which makes it buggy when patch is created via YAML in ICSP, or dynamic patch in CPEntity).

This diff added support to incomplete type in MapPatch so for structs that contains map to itself, we don't need AssignOnlyPatch anymore.

Reviewed By: thedavekwon

Differential Revision: D52427991

fbshipit-source-id: d1b80eb6667f04fe88c5b9e413831a44b16f6d45