RFC: Rearrange how error types are exposed in services&clients crates
Summary:
## Background

If you have a Thrift library called `validation_thrift` containing a service called `ValidationService` containing a method called `validate`, the Thrift compiler generates an error type called `ValidateExn` that whoever implements a `ValidationService` server is supposed to return, and an error type called `ValidateError` that callers of the `ValidationService` client receive.

Both error types are enums with a variant for each exception that the Thrift service declares can be thrown by the `validate` method.

The server-side error and client-side error are different types because the client one contains 1 more variant: for errors that occurred in the plumbing between the server and client. A server is not supposed to be able to produce that variant. For example if a transport error occurs, such as socket disconnected, the error variant received by the client in that situation is not something that should have been able to be produced directly by the server.

## Before

The error types are exposed in the following locations:

- `validation_thrift::errors::validation_service::ValidateError`
- `validation_thrift::validation_service::ValidateError`
- `validation_thrift::services::validation_service::ValidateExn`
- `validation_thrift_clients::errors::validation_service::ValidateError`
- `validation_thrift_services::errors::validation_service::ValidateExn`

Critiques:

- Notice how every one of these involves typing out the snakecased service name. This makes paths needlessly verbose because the vast majority of Thrift libraries contain only at most a single service, so the service name in these paths is mostly not providing value. Among the Thrift libraries that contain more than a single service, the vast majority do not repeat the same method name in multiple services in the same library.

- Top-level `errors` and `services` modules are not well named. One must remember that they both contain errors, with `errors` containing only client-side errors and `services` containing the server-side errors. `services` never contained service traits. Those used to be in a module called `server`, adding to the confusion, until being deleted by D54468159.

- Top-level `validation_service` arbitrarily disambiguates to containing only client error types, not server error types. There isn't a principled reason why this would be, other than maybe client code is written more often than server code.

## After

- `validation_thrift_clients::errors::ValidateError`
- `validation_thrift_clients::errors::validation_service::ValidateError`
- `validation_thrift_services::errors::ValidateExn`
- `validation_thrift_services::errors::validation_service::ValidateExn`.
- `validation_thrift`: not exposed (needs codemod)

Just the method name is enough to give unique error names most of the time. For the rare case where a Thrift library contains multiple services with colliding method names, one can include the service name in the path to disambiguate.

This is done by having `errors` do `pub use validation_service::*` for each service. Type names that collide do not get re-exported by the glob.

Reviewed By: emersonford

Differential Revision: D54514896

fbshipit-source-id: 915d9123404980f79f1bad3da23ccc11157a7a2b