Fix bug when enabling lazy deserialization with thrift.Box
Summary:
# What is serializedSize(...)?

For thrift struct, we have `foo.serializedSize(Protocol*)` method, It's an estimation of the actual size.

When we do RPC, we will first call `serializedSize` to get the size estimation and allocate memory, so that we can improve the performance by only allocating memory once (If we don't pre-allocate memory, IOBufQueue will allocate memory multiple times when we increase the buffer size).

However, serializedSize is not 100% accurate. e.g., if we have a `list<i32>`, since we use varint, we don't know the actual size for each integer. In this case, we assume each integer needs 5 bytes since that's the upperbound.

There is also serializedSizeZC (ZC stands for zero copy). The difference is that if the field is IOBuf, ZC needs 0 bytes (since it can just link to the existing IOBuf without copying the memory). However, without ZC, it will will be the length of underlying IOBuf. (this is useful when you serialize to `std::string` where you can link to the existing IOBuf).

RPC actually use serializedSizeZC version

# Problem

Currently in serializedSize(...), for thrift.box field, if it has undeserialized iobuf, or the pointer is not nullptr, we will

```
if (has undeserialized iobuf || this->member != nullptr) {
  xfer += protocol_methods::serialized_size(*this->member);
}
```

This is incorrect. We should mimic the existing serialization code and change to

```
if (has undeserialized iobuf || this->member != nullptr) {

  // If iobuf is untouched, we can use its length to estimate the size.
  if (undeserialized iobuf is untouched) {
    xfer += length of iobuf;
    goto finished;
  }

  __fbthrift_read_field_member(); // force deserialize member to make sure `this->member != nullptr`.

  protocol_methods::serialized_size(*this->member);
  finished:
}
```

Reviewed By: iahs

Differential Revision: D54760367

fbshipit-source-id: 704823ad4e1ef0d8dfddc679b83ee0b0202977b5